requirements.txt
aiohttp==3.10.5
alembic==1.13.2
annotated-types==0.7.0
asyncpg==0.29.0
bcrypt==4.0.1
black==24.8.0
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
fastapi-versioning==0.10.0
frozenlist==1.4.1
gunicorn==23.0.0
packaging==24.1
passlib==1.7.4
platformdirs==4.2.2
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
websockets==13.0
yarl==1.9.4
